+++ src/lib/Scenes/AuctionResult/__tests__/AuctionResult-tests.tsx
@@ -1,9 +1,10 @@
 import { FancyModalHeader } from "lib/Components/FancyModal/FancyModalHeader"
 import { extractText } from "lib/tests/extractText"
+import { mockEnvironmentPayload } from "lib/tests/mockEnvironmentPayload"
 import { renderWithWrappers } from "lib/tests/renderWithWrappers"
 import React from "react"
 import { ScrollView } from "react-native"
-import { createMockEnvironment, MockPayloadGenerator } from "relay-test-utils"
+import { createMockEnvironment } from "relay-test-utils"
 import { AuctionResultQueryRenderer } from "../AuctionResult"
 
 jest.unmock("react-relay")
@@ -19,7 +20,7 @@ describe("AuctionResult", () => {
         environment={mockEnvironment}
       />
     )
-    mockEnvironment.mock.resolveMostRecentOperation((operation) => MockPayloadGenerator.generate(operation, mockedData))
+    mockEnvironmentPayload(mockEnvironment, mockedData)
     return tree
   }
 
@@ -64,48 +65,3 @@ describe("AuctionResult", () => {
     expect(tree.root.findAllByProps({ testID: "ratio" })).toHaveLength(0)
   })
 })
-
-//   it("renders auction is closed when an auction has passed", () => {
-//     const tree = renderWithWrappers(<TestRenderer />)
-
-//     mockEnvironment.mock.resolveMostRecentOperation((operation) =>
-//       MockPayloadGenerator.generate(operation, {
-//         Sale: () => ({
-//           endAt: moment().subtract(1, "day").toISOString(),
-//           startAt: "2020-09-01T15:00:00",
-//           timeZone: "Europe/Berlin",
-//           coverImage: {
-//             url: "cover image url",
-//           },
-//           name: "sale name",
-//           liveStartAt: "2020-09-01T15:00:00",
-//           internalID: "the-sale-internal",
-//         }),
-//       })
-//     )
-
-//     expect(extractText(tree.root.findAllByType(OpaqueImageView)[0])).toBe("Auction closed")
-//   })
-
-//   it("does not render auction is closed when an auction is still active", () => {
-//     const tree = renderWithWrappers(<TestRenderer />)
-
-//     mockEnvironment.mock.resolveMostRecentOperation((operation) =>
-//       MockPayloadGenerator.generate(operation, {
-//         Sale: () => ({
-//           endAt: moment().add(1, "day").toISOString(),
-//           startAt: "2020-09-01T15:00:00",
-//           timeZone: "Europe/Berlin",
-//           coverImage: {
-//             url: "cover image url",
-//           },
-//           name: "sale name",
-//           liveStartAt: "2020-09-01T15:00:00",
-//           internalID: "the-sale-internal",
-//         }),
-//       })
-//     )
-
-//     expect(extractText(tree.root.findAllByType(OpaqueImageView)[0])).not.toContain("Auction closed")
-//   })
-// })
